Complete 4.23

+++ osa4/blogilista/controllers/blogs.js
@@ -1,16 +1,8 @@
 const blogsRouter = require('express').Router()
 const { Error } = require('mongoose')
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')){
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+const config = require('../utils/config')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
@@ -19,33 +11,44 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, config.TOKENSECRET)
+    if (!decodedToken.id){
+      return response.status(401).json({ error: 'invalid token!' })
+    }
+
+    const blogToDelete = await Blog.findById(request.params.id)
+    console.log('Blogi', blogToDelete)
+    console.log('Token', decodedToken.id)
+    if (decodedToken.id.toString() === blogToDelete.user.toString()){
+      await Blog.findByIdAndDelete(request.params.id)
+      return response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'invalid user!' })
+    }
   }
   catch(error) {next(error)}
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
-
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-
-  if (!decodedToken.id){
-    return response.status(401).json({ error: 'invalid token!' })
-  }
-
-  const user = await User.findById(decodedToken.id)
-
-  const blog = new Blog({
-    id: body.id,
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user._id
-  })
-
   try{
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, config.TOKENSECRET)
+
+    if (!decodedToken.id){
+      return response.status(401).json({ error: 'invalid token!' })
+    }
+
+    const user = request.user
+    console.log('Käyttäjä', user)
+
+    const blog = new Blog({
+      id: body.id,
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user.id
+    })
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
